fix(washboard-ui): fast refresh for settings-context to resolve esling warning
